refactor: 함수 분리 및 리팩토링
